Added SVGIsReactComponent prop to allow React child
I have a react component that renders SVG, and was looking to add
pan/zoom to it.

This component is perfect, expect that it checks that it gets
passed a single child of type `svg`.

I have added a prop named `SVGIsReactComponent` that has a default
value of `false`. Setting it to `true` will skip the props.children
check and simply render the wrapped component.

It works fine for my use-case — the wrapped component not only
displays correctly, I can also still interact with it — and this
should have no impact whatsoever if the `SVGIsReactComponent`
prop is not set (or set to `false`).